ignore ost/mdt's that are connections to other lustre servers
we're only interested in the i/o rates between clients and servers and
not from server to server.

firstly reading zero's is distinguished from non-existent data. then
the identities of server-to-server connections is inferred from the
patterns read from the ost/mdt stats files. server-server ost/mdts are
then filtered out and not stored.

obviously this form of automatic detection may not be very robust to
future lustre changes. one alternative would be to explicitly specify
the ips of all the lustre servers but that complicates installation.